+++ frontend/src/pages/today-habit/TodayHabit.jsx
@@ -1,6 +1,6 @@
 import HabitModal from "@today-habit/HabitModal";
-import React, { useEffect, useState } from "react";
-import { Link, useParams } from "react-router-dom";
+import React, { useState } from "react";
+import { useParams } from "react-router-dom";
 import styles from "@today-habit/TodayHabit.module.css";
 import Habits from "@today-habit/Habits";
 import StudyNavbar from "@components/study-navbar/StudyNavbar";
